+++ cmd/wallet.go
@@ -76,7 +76,6 @@ func init() {
 	walletCmd.AddCommand(walletTLSCertCmd())
 	walletCmd.AddCommand(walletInfoCmd())
 	walletCmd.AddCommand(walletInstancesCmd())
-	walletCmd.AddCommand(walletConformanceCmd())
 
 	// Deprecated aliases (hidden from help)
 	presentAlias := &cobra.Command{